feat(charcount): use double instead of long

+++ ch1/src/charcount.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 int main() {
-    long nc;
+    double nc;
     nc = 0;
     while (getchar() != EOF) {
         ++nc;
     }
-    printf("%ld\n", nc);
+    printf("%.0f\n", nc);
     return 0;
 }
